% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{calc_consensus}
\alias{calc_consensus}
\title{Calculate consensus matrix2}
\usage{
calc_consensus(matrix, k)
}
\arguments{
\item{matrix}{a matrix containing clustering solutions in columns}

\item{k}{number of clusters}
}
\value{
consensus matrix
}
\description{
For each clustering solution a binary
similarity matrix is constructed from the corresponding cell labels:
if two cells belong to the same cluster, their similarity is 1, otherwise
the similarity is 0. A consensus matrix is calculated by averaging all
similarity matrices.
}
