% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate a distance matrix}
\usage{
calculate_distance(d, method)
}
\arguments{
\item{d}{Expression matrix with rows as genes and columns as cells}

\item{method}{Distance metrics: "spearman", "pearson", "euclidean", "maximum",
"manhattan", "canberra", "binary" or "minkowski"}
}
\value{
A distance matrix
}
\description{
Calculate a distance between column vectors of the input dataset using a
specified distance metrics.
}

