% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{calculate_omics_consensus}
\alias{calculate_omics_consensus}
\title{Multiply consensus matrices from all omics}
\usage{
calculate_omics_consensus(matrices)
}
\arguments{
\item{matrices}{consensus matrices from each omic}
}
\value{
consensus of all omics
}
\description{
For each omics, take consensus matrix and multiply it with the consensus matrix of the other omics. Take only cells that are common 
between all omics.
}
