% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{cell_filter}
\alias{cell_filter}
\title{Cell filter}
\usage{
cell_filter(data, cell.filter.genes)
}
\arguments{
\item{data}{Input expression matrix}

\item{cell.filter.genes}{number of genes that have to expressed in each cell}
}
\value{
Filtered expression matrix
}
\description{
More than cell.filter.genes have to be expressed in each cell
}

