% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{consensus_clustering}
\alias{consensus_clustering}
\title{Consensus clustering}
\usage{
consensus_clustering(clusts)
}
\arguments{
\item{clusts}{a list clustering labels (separated by a space)}
}
\value{
Consensus matrix
}
\description{
Caclulate consensus clustering by using Cluster-based similarity
partitioning algorithm (CSPA)
}

