% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{create_sceset}
\alias{create_sceset}
\title{A wrapper to quickly create an SCESet from a matrix}
\usage{
create_sceset(d)
}
\arguments{
\item{d}{expression matrix}
}
\value{
object of SCESet class based on d
}
\description{
If an input matrix has colnames then the phenoData table of the output
object will be populated with these colnames as 'cell_id' column.
}
\details{
If an input matrix does not have colnames the phenoData table of the output
object will be populated with 'Cell_1', 'Cell_2', etc. as 'cell_id' column.
}

