% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ED2}
\alias{ED2}
\title{Compute Euclidean distance matrix by columns}
\usage{
ED2(x)
}
\arguments{
\item{x}{A numeric matrix.}
}
\description{
Used in sc3-funcs.R distance matrix calculation
and within the consensus clustering.
}
