% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estkTW.R
\name{estkTW}
\alias{estkTW}
\title{Estimate the optimal k for k-means clustering}
\usage{
estkTW(dataset, gene.filter = TRUE, gene.filter.fraction = 0.06,
  gene.reads.rare = 2, gene.reads.ubiq = 0, log.scale = TRUE)
}
\arguments{
\item{dataset}{input expression matrix. The expression
matrix must contain both colnames (cell IDs) and rownames (gene IDs).}

\item{gene.filter}{defines whether to perform gene filtering or not. Boolean,
default is TRUE.}

\item{gene.filter.fraction}{fraction of cells (1 - X/100), default is 0.06.
The gene filter removes genes that are either expressed or absent
(expression value is less than 2) in at least X % of cells.
The motivation for the gene filter is that ubiquitous and rare genes most
often are not informative for the clustering.}

\item{gene.reads.rare}{expression value threshold for genes that are expressed in
less than gene.filter.fraction*N cells (rare genes)}

\item{gene.reads.ubiq}{expression value threshold for genes that are expressed in
more than (1-fraction)*N cells (ubiquitous genes)}

\item{log.scale}{defines whether to perform log2 scaling or not. Boolean,
default is TRUE.}
}
\description{
The procedure filters and log-transforms the expression matrix before 
finding the eigenvalues of the sample covariance matrix. It will then return 
the number of significant eigenvalues according to the Tracy-Widom test.
}

