% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{FindSimilarities}
\alias{FindSimilarities}
\title{Creates a binary similarity matrix for every clustering created using the consensus algorithm and clustering with k-means after that.}
\usage{
FindSimilarities(v)
}
\arguments{
\item{v}{vector containing clustering results}
}
\value{
binary similarity matrix
}
\description{
Finds whether 2 cells belong to the same cluster in a given clustering result.
}
