% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{gene_filter}
\alias{gene_filter}
\title{Gene filter}
\usage{
gene_filter(data, fraction)
}
\arguments{
\item{data}{Input expression matrix}

\item{fraction}{threshold defining a fraction cells for each gene}
}
\value{
Filtered expression matrix
}
\description{
Filter genes that would not contribute to clustering, because they are either
expressed or not expressed in almost all cells.
}

