% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{gene_filter1}
\alias{gene_filter1}
\title{Gene filter}
\usage{
gene_filter1(d, min.cells, max.cells, min.reads)
}
\arguments{
\item{d}{Expression matrix with rows as genes and columns as cells}

\item{min.cells}{Minimum number of cells in which a given gene is expressed
(obtained from filter_params())}

\item{max.cells}{Maximum number of cells in which a given gene is expressed
(obtained from filter_params())}

\item{min.reads}{Minimum number of reads per gene per cell
(obtained from filter_params())}
}
\value{
Filtered expression matrix in which only genes that are expressed in
more than \code{min.cells} with more than \code{min.reads} reads and also are
expressed in less than [total number of cells - \code{max.cells}].
}
\description{
Filter genes that would not contribute to clustering, because they are either
expressed or not expressed in almost all cells.
}

