% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{get_auroc}
\alias{get_auroc}
\title{Calculate the area under the ROC curve for a given gene.}
\usage{
get_auroc(gene, labels)
}
\arguments{
\item{gene}{expression data of a given gene}

\item{labels}{cell labels correspodning to the expression values of the gene}
}
\description{
For a given gene a binary classifier is constructed 
based on the mean cluster expression values (these are calculated using the
cell labels). The classifier prediction 
is then calculated using the gene expression ranks. The area under the 
receiver operating characteristic (ROC) curve is used to quantify the accuracy 
of the prediction. A \code{p-value} is assigned to each gene by using the Wilcoxon 
signed rank test.
}
