% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{get_biolgy}
\alias{get_biolgy}
\title{Wrapper for calculating biological properties}
\usage{
get_biolgy(dataset, labels, regime)
}
\arguments{
\item{dataset}{expression matrix}

\item{labels}{cell labels corresponding clusters}

\item{regime}{defines what biological analysis to perform. "marker" for
marker genes, "de" for differentiall expressed genes and "outl" for outlier
cells}
}
\value{
results of either
}
\description{
Wrapper for calculating biological properties
}

