% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_common_cells}
\alias{get_common_cells}
\title{Get common cells from all matrices}
\usage{
get_common_cells(matrices)
}
\arguments{
\item{matrices}{consensus matrices from each omic}
}
\value{
list of matrices with the same number of cells
}
\description{
For each omics, take consensus matrix and multiply it with the consensus matrix of the other omics. Take only cells that are common 
between all omics.
}
