% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_consensus_matrix}
\alias{get_consensus_matrix}
\title{Calculate consensus matrix}
\usage{
get_consensus_matrix(clusts, k)
}
\arguments{
\item{clusts}{a matrix containing clustering solutions in columns}

\item{k}{number of clusters}
}
\value{
consensus matrix
}
\description{
Calculate consensus matrix
}
