% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{get_data}
\alias{get_data}
\title{Import expression matrix}
\usage{
get_data(name)
}
\arguments{
\item{name}{either treutlein or a name of a text file}
}
\value{
Expression matrix
}
\description{
Import either treutlein dataset or a dataset defined by a user
}

