% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-funcs.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Find marker genes}
\usage{
get_marker_genes(dataset, labels, auroc.threshold = 0.85, p.val = 0.01)
}
\arguments{
\item{dataset}{expression matrix}

\item{labels}{cell labels corresponding to the columns of the expression matrix.
Labels must be integers or character integers, e.g. 1, 2, 3 or "1", "2", "3" ect.}

\item{auroc.threshold}{area under the ROC curve threshold, by default it is
0.85. Values close to 0.5 will include very weak marker genes, values close
to 1 will only include very strong marker genes.}

\item{p.val}{p-value threshold, by default it is 0.01}
}
\value{
data.frame containing the marker genes
}
\description{
If the cell labels are available this functions allows a user to calculate
marker genes manually.
}
\examples{
d <- get_marker_genes(treutlein, colnames(treutlein))
head(d)

}

