% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{markers_for_heatmap}
\alias{markers_for_heatmap}
\title{Reorder and subset gene markers for plotting on a heatmap}
\usage{
markers_for_heatmap(markers)
}
\arguments{
\item{markers}{a \code{data.frame} object with the following colnames:
\code{moSC3_k_markers_clusts}, \code{moSC3_k_markers_auroc}, \code{moSC3_k_markers_padj}.}
}
\description{
Reorders the rows of the input data.frame based on the \code{moSC3_k_markers_clusts}
column and also keeps only the top 10 genes for each value of \code{moSC3_k_markers_clusts}.
}
