% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{moSC3_calc_dists}
\alias{moSC3_calc_dists}
\alias{moSC3_calc_dists.SingleCellExperiment}
\alias{moSC3_calc_dists,}
\alias{moSC3_calc_dists,SingleCellExperiment-method}
\title{Calculate distances between the cells.}
\usage{
moSC3_calc_dists.SingleCellExperiment(object)

\S4method{moSC3_calc_dists}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function calculates distances between the cells. It
creates and populates the following items of the \code{moSC3} slot of the \code{metadata(object)}:
\itemize{
  \item \code{distances} - contains a list of distance matrices corresponding to
  Euclidean, Pearson and Spearman distances.
}
}
