% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{moSC3_estimate_k}
\alias{moSC3_estimate_k}
\alias{moSC3_estimate_k.list}
\alias{moSC3_estimate_k.list-method}
\alias{moSC3_estimate_k,list-method}
\title{Estimate the optimal number of cluster \code{k} for all omics}
\usage{
moSC3_estimate_k.list(object)

\S4method{moSC3_estimate_k}{list}(object)
}
\arguments{
\item{object}{a list of \code{SingleCellExperiment} objects}
}
\value{
an estimated value of k
}
\description{
Uses Tracy-Widom theory on random matrices to estimate the optimal number of
clusters \code{k}. It creates and populates the \code{k_estimation} item of the
\code{moSC3} slot of the \code{metadata(object)}.
}
