% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{moSC3_estimate_k_single_omic}
\alias{moSC3_estimate_k_single_omic}
\alias{moSC3_estimate_k_single_omic.SingleCellExperiment}
\alias{moSC3_estimate_k_single_omic,}
\alias{moSC3_estimate_k_single_omic,SingleCellExperiment-method}
\title{Estimate the optimal number of cluster \code{k} for one omic}
\usage{
moSC3_estimate_k_single_omic.SingleCellExperiment(object)

\S4method{moSC3_estimate_k_single_omic}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
an estimated value of k
}
\description{
Uses Tracy-Widom theory on random matrices to estimate the optimal number of
clusters \code{k}. It creates and populates the \code{k_estimation} item of the
\code{moSC3} slot of the \code{metadata(object)}.
}
