% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{moSC3_export_results_xls}
\alias{moSC3_export_results_xls}
\alias{moSC3_export_results_xls.SingleCellExperiment}
\alias{moSC3_export_results_xls,SingleCellExperiment-method}
\title{Write \code{moSC3} results to Excel file}
\usage{
moSC3_export_results_xls.SingleCellExperiment(object, filename)

\S4method{moSC3_export_results_xls}{SingleCellExperiment}(object,
  filename = "moSC3_results.xls")
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{filename}{name of the excel file, to which the results will be written}
}
\description{
This function writes all \code{moSC3} results to an excel file.
}
