% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{moSC3_interactive}
\alias{moSC3_interactive}
\alias{moSC3_interactive.SingleCellExperiment}
\alias{moSC3_interactive,}
\alias{moSC3_interactive,SingleCellExperiment-method}
\title{Opens \code{moSC3} results in an interactive session in a web browser.}
\usage{
moSC3_interactive.SingleCellExperiment(object)

\S4method{moSC3_interactive}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize
all precomputed clusterings.
}
\description{
Runs interactive \code{shiny} session of \code{moSC3} based on precomputed clusterings.
}
