% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{moSC3_plot_cluster_stability}
\alias{moSC3_plot_cluster_stability}
\alias{moSC3_plot_cluster_stability.SingleCellExperiment}
\alias{moSC3_plot_cluster_stability,}
\alias{moSC3_plot_cluster_stability,SingleCellExperiment-method}
\title{Plot stability of the clusters}
\usage{
moSC3_plot_cluster_stability.SingleCellExperiment(object, k)

\S4method{moSC3_plot_cluster_stability}{SingleCellExperiment}(object, k)
}
\arguments{
\item{object}{an object of 'SingleCellExperiment' class}

\item{k}{number of clusters}
}
\description{
Stability index shows how stable each cluster is accross the selected 
range of ks. The stability index varies between 0 and 1, where 1 means that 
the same cluster appears in every solution for different k.
}
