% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{moSC3_plot_de_genes}
\alias{moSC3_plot_de_genes}
\alias{moSC3_plot_de_genes.SingleCellExperiment}
\alias{moSC3_plot_de_genes,}
\alias{moSC3_plot_de_genes,SingleCellExperiment-method}
\title{Plot expression of DE genes of the clusters identified by \code{moSC3} as a heatmap}
\usage{
moSC3_plot_de_genes.SingleCellExperiment(object, k, p.val, show_pdata)

\S4method{moSC3_plot_de_genes}{SingleCellExperiment}(object, k,
  p.val = 0.01, show_pdata = NULL)
}
\arguments{
\item{object}{an object of 'SingleCellExperiment' class}

\item{k}{number of clusters}

\item{p.val}{significance threshold used for the DE genes}

\item{show_pdata}{a vector of colnames of the pData(object) table. Default is NULL.
If not NULL will add pData annotations to the columns of the output matrix}
}
\description{
\code{moSC3} plots gene expression profiles of the 50 genes with the lowest p-values.
}
