% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult}
\alias{mult}
\title{More efficient way to pre and post multiply by diagonal matrix D}
\usage{
mult(D, x, dim)
}
\arguments{
\item{D}{Diagonal numeric matrix.}

\item{x}{Numeric matrix with the same dimensions as D.}

\item{dim}{Number of rows of D.}
}
\description{
If D is a diagonal matrix and A another matrix with the same
dimensions, this procedure is a more efficient way to compute
D %*% A %*% D.
}

