% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC3class.R
\name{newSC3}
\alias{newSC3}
\title{Create a new object of SC3 class}
\usage{
newSC3(object)
}
\arguments{
\item{object}{an object of `SingleCellExperiment` class}
}
\value{
an object of `SC3` class
}
\description{
Create a new object of SC3 class
}
