% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{norm_laplacian}
\alias{norm_laplacian}
\title{Laplacian}
\usage{
norm_laplacian(x, tau)
}
\arguments{
\item{x}{Adjacency/distance matrix}

\item{tau}{Regularization term}
}
\value{
Laplacian of the adjacency/distance matrix
}
\description{
Calculate Laplacian of an input distance matrix
}

