% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{organise_de_genes}
\alias{organise_de_genes}
\title{Get differentiall expressed genes from an object of \code{SCESet} class}
\usage{
organise_de_genes(object, k, p_val)
}
\arguments{
\item{object}{an objec of \code{SCESet} class}

\item{k}{number of cluster}

\item{p_val}{p-value threshold}
}
\description{
This functions returns all marker gene columns from the \code{phenoData} slot 
of the input object corresponding to the number of clusters \code{k}. Additionally,
it rearranges genes by the cluster index and order them by the area under the 
ROC curve value inside of each cluster.
}
