% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{sc3.interactive.example}
\alias{sc3.interactive.example}
\title{Data used for sc3_interactive() example}
\format{An object of class \code{list} of length 11.}
\usage{
sc3.interactive.example
}
\value{
blah blah
}
\description{
Data used for sc3_interactive() example
}
\keyword{datasets}

