% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3.SCESet}
\alias{sc3}
\alias{sc3,SCESet-method}
\alias{sc3.SCESet}
\title{Run all steps of SC3 in one go}
\usage{
sc3.SCESet(object, exprs_values = "exprs", gene_filter = TRUE,
  pct_dropout_min = 10, pct_dropout_max = 90, log.scale = FALSE,
  d.region.min = 0.04, d.region.max = 0.07, svm.num.cells = NULL,
  svm.train.inds = NULL, svm.max = 5000, n.cores = NULL, ks = NULL,
  k.means.nstart = NULL, k.means.iter.max = 1e+09, biology = TRUE,
  seed = 1)

\S4method{sc3}{SCESet}(object, exprs_values = "exprs", gene_filter = TRUE,
  pct_dropout_min = 10, pct_dropout_max = 90, gene.reads.ubiq = 0,
  log.scale = FALSE, d.region.min = 0.04, d.region.max = 0.07,
  svm.num.cells = NULL, svm.train.inds = NULL, svm.max = 5000,
  n.cores = NULL, ks = NULL, k.means.nstart = NULL,
  k.means.iter.max = 1e+09, biology = TRUE, seed = 1)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{exprs_values}{character string 
indicating which values should be used
as the expression values for SC3 clustering. Valid arguments are \code{'tpm'}
(default; transcripts per million), \code{'norm_tpm'} (normalised TPM
values), \code{'fpkm'} (FPKM values), \code{'norm_fpkm'} (normalised FPKM
values), \code{'counts'} (counts for each feature), \code{'norm_counts'},
\code{'cpm'} (counts-per-million), \code{'norm_cpm'} (normalised
counts-per-million), \code{'exprs'} (whatever is in the \code{'exprs'} slot
of the \code{SCESet} object; default), \code{'norm_exprs'} (normalised
expression values) or \code{'stand_exprs'} (standardised expression values)
or any other named element of the \code{assayData} slot of the \code{SCESet}
object that can be accessed with the \code{get_exprs} function.}

\item{gene_filter}{a boolen variable which defines whether to perform gene 
filtering before SC3 clustering.}

\item{pct_dropout_min}{genes with percent of dropouts smaller than 
pct_dropout_min are filtered out before clustering}

\item{pct_dropout_max}{genes with percent of dropouts larger than 
pct_dropout_max are filtered out before clustering}

\item{log.scale}{a boolean variable which defines whether to perform log2 
scaling before SC3 clustering. Default is TRUE.}

\item{d.region.min}{defines the minimum number of eigenvectors used for 
kmeans clustering as a fraction of the total number of cells. Default is 0.04.}

\item{d.region.max}{defines the maximum number of eigenvectors used for 
kmeans clustering as a fraction of the total number of cells. Default is 0.07.}

\item{svm.num.cells}{number of randomly selected training cells to be used 
for SVM prediction. The default is NULL.}

\item{svm.train.inds}{a numeric vector defining indeces of training cells 
that should be used for SVM training. The default is NULL.}

\item{svm.max}{define the maximum number of cells below which SVM is not run.}

\item{n.cores}{defines the number of cores to be used on the user's machine.}

\item{ks}{a range of the number of clusters k used for SC3 clustering.
Can also be a single integer.}

\item{k.means.nstart}{nstart parameter used by kmeans() function. Default is 
1000 for up to 2000 cells and 50 for more than 2000 cells.}

\item{k.means.iter.max}{iter.max parameter passed to \link[stats]{kmeans} 
function. Default is 1e+09.}

\item{biology}{boolean variable, defines whether to comput DE genes, marker 
genes and cell outliers}

\item{seed}{sets seed for the random number generator.
Can be used to check the stability of clustering results: if the results are 
the same after changing the seed several time, then the clustering solution 
is stable.}
}
\value{
an object of 'SCESet' class
}
\description{
This function is a wrapper that executes all steps of SC3 analysis in one go.
}

