% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_calc_biology.SCESet}
\alias{sc3_calc_biology}
\alias{sc3_calc_biology,SCESet-method}
\alias{sc3_calc_biology.SCESet}
\title{Calculate DE genes, marker genes and cell outliers.}
\usage{
sc3_calc_biology.SCESet(object)

\S4method{sc3_calc_biology}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function calculates DE genes, marker genes and cell outliers based on 
the consensus clusterings
contained in the consensus item of the object@sc3 slot. It then
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item biology - contains lists of DE genes, marker genes and 
  cell outliers data frames.
}
}

