% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_calc_consens.SCESet}
\alias{sc3_calc_consens}
\alias{sc3_calc_consens,SCESet-method}
\alias{sc3_calc_consens.SCESet}
\title{Calculate consensus matrix.}
\usage{
sc3_calc_consens.SCESet(object)

\S4method{sc3_calc_consens}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function calculates consensus matrices based on the clustering solutions
contained in the sc3_kmeans item of the object@sc3 slot. It then
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item consensus - contains a list of consensus matrices. In addition 
  to consensus matrices it also contains the Silhouette
  indeces of the clusters and original cell labels corresponding to the clusters.
}
}

