% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{sc3_calc_dists}
\alias{sc3_calc_dists}
\alias{sc3_calc_dists.SingleCellExperiment}
\alias{sc3_calc_dists,}
\alias{sc3_calc_dists,SingleCellExperiment-method}
\title{Calculate distances between the cells.}
\usage{
sc3_calc_dists.SingleCellExperiment(object, BPPARAM)

\S4method{sc3_calc_dists}{SingleCellExperiment}(object, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}} object specifying a type of parallelism and allocated resources,
including a seed for random number generator.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function calculates distances between the cells. It
creates and populates the following items of the \code{sc3} slot of the \code{metadata(object)}:
\itemize{
  \item \code{distances} - contains a list of distance matrices corresponding to
  Euclidean, Pearson and Spearman distances.
}
}
