% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_calc_dists.SCESet}
\alias{sc3_calc_dists}
\alias{sc3_calc_dists,SCESet-method}
\alias{sc3_calc_dists.SCESet}
\title{Calculate distances between the cells.}
\usage{
sc3_calc_dists.SCESet(object)

\S4method{sc3_calc_dists}{SCESet}(object)
}
\arguments{
\item{object}{an object of "SCESet" class}
}
\value{
an object of "SCESet" class
}
\description{
This function calculates distances between the cells contained in 
the sc3_processed_dataset item of the object@consensus slot. It then
creates and populates the following items of the object@consensus slot:
\itemize{
  \item sc3_distances - contains a list of distance matrices corresponding to
  Euclidean, Pearson and Spearman distances.
}
}

