% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_calc_transfs.SCESet}
\alias{sc3_calc_transfs}
\alias{sc3_calc_transfs,SCESet-method}
\alias{sc3_calc_transfs.SCESet}
\title{Calculate transformations of the distance matrices.}
\usage{
sc3_calc_transfs.SCESet(object)

\S4method{sc3_calc_transfs}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function calculates transforamtions of the distance matrices contained in 
the sc3_distances item of the object@sc3 slot. It then
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item transformations - contains a list of transformations of the 
  distance matrices corresponding to PCA and graph Laplacian transformations.
}
}

