% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_estimate_k}
\alias{sc3_estimate_k}
\alias{sc3_estimate_k.SingleCellExperiment}
\alias{sc3_estimate_k,SingleCellExperiment-method}
\alias{sc3_estimate_k,SingleCellExperiment-method}
\alias{sc3_estimate_k}
\title{Estimate the optimal number of cluster \code{k} for a scRNA-Seq expression matrix}
\usage{
sc3_estimate_k.SingleCellExperiment(object)

\S4method{sc3_estimate_k}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}
}
\value{
an estimated value of k
}
\description{
Uses Tracy-Widom theory on random matrices to estimate the optimal number of
clusters \code{k}. It creates and populates the \code{k_estimation} item of the
\code{sc3} slot of the \code{metadata(object)}.
}
