% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_estimate_k.SCESet}
\alias{sc3_estimate_k}
\alias{sc3_estimate_k,SCESet-method}
\alias{sc3_estimate_k.SCESet}
\title{Estimate the optimal k for k-means clustering}
\usage{
sc3_estimate_k.SCESet(object)

\S4method{sc3_estimate_k}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an estimated value of k
}
\description{
Uses Tracy-Widom theory on random matrices to estimate the optimal number of
clusters k. Using the function \code{\link{estkTW}} to perform the estimation. 
It creates and populates the following items of the `sc3` slot:
\itemize{
  \item k_prediction - contains the estimated value of `k`.
}
}

