% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{sc3_kmeans}
\alias{sc3_kmeans}
\alias{sc3_kmeans.SingleCellExperiment}
\alias{sc3_kmeans,}
\alias{sc3_kmeans,SingleCellExperiment-method}
\title{\code{kmeans} clustering of cells.}
\usage{
sc3_kmeans.SingleCellExperiment(object, ks, BPPARAM)

\S4method{sc3_kmeans}{SingleCellExperiment}(object, ks = NULL, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{ks}{a continuous range of integers - the number of clusters \code{k} to be used for SC3 clustering.
Can also be a single integer.}

\item{BPPARAM}{a \code{\link[BiocParallel]{BiocParallelParam}} object specifying a type of parallelism and allocated resources,
including a seed for random number generator.}
}
\value{
an object of \code{SingleCellExperiment} class
}
\description{
This function performs \code{\link[stats]{kmeans}} clustering of the matrices
contained in the \code{transformations} item of the \code{sc3} slot of the \code{metadata(object)}. It then
creates and populates the following items of the \code{sc3} slot:
\itemize{
  \item \code{kmeans} - contains a list of kmeans clusterings.
}
}
