% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_cell_outliers.SCESet}
\alias{sc3_plot_cell_outliers}
\alias{sc3_plot_cell_outliers,SCESet-method}
\alias{sc3_plot_cell_outliers.SCESet}
\title{Plot cell outliers}
\usage{
sc3_plot_cell_outliers.SCESet(object, k)

\S4method{sc3_plot_cell_outliers}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}
}
\description{
Outlier cells in each cluster are detected using robust distances, 
calculated using the minimum covariance determinant (MCD). 
The outlier score shows how different a cell is from all other cells in the 
cluster and it is defined as the differences between the square root of the 
robust distance and the square root of the 99.99% quantile of the 
Chi-squared distribution.
}

