% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_cluster_stability}
\alias{sc3_plot_cluster_stability}
\alias{sc3_plot_cluster_stability,}
\alias{sc3_plot_cluster_stability,SCESet-method}
\alias{sc3_plot_cluster_stability.SCESet}
\title{Plot stability of the clusters}
\usage{
sc3_plot_cluster_stability.SCESet(object, k)

\S4method{sc3_plot_cluster_stability}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}
}
\description{
Stability index shows how stable each cluster is accross the selected 
range of ks. The stability index varies between 0 and 1, where 1 means that 
the same cluster appears in every solution for different k.
}

