% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_consensus.SCESet}
\alias{sc3_plot_consensus}
\alias{sc3_plot_consensus,SCESet-method}
\alias{sc3_plot_consensus.SCESet}
\title{Plot consensus matrix as a heatmap}
\usage{
sc3_plot_consensus.SCESet(object, k)

\S4method{sc3_plot_consensus}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}
}
\description{
The consensus matrix is a NxN 
matrix, where N is the number of cells.
It represents similarity between the cells based 
on the averaging of clustering results from all 
combinations of clustering parameters. Similarity 0 
(blue) means that the two cells are always assigned to different clusters. 
In contrast, similarity 1 (red) means that the two cells are always assigned 
to the same cluster. The consensus matrix is clustered by hierarchical 
clustering and has a diagonal-block structure. Intuitively, the perfect 
clustering is achieved when all diagonal blocks are completely red 
and all off-diagonal elements are completely blue.
}

