% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_de_genes}
\alias{sc3_plot_de_genes}
\alias{sc3_plot_de_genes,}
\alias{sc3_plot_de_genes,SCESet-method}
\alias{sc3_plot_de_genes.SCESet}
\title{Plot expression of DE genes of the clusters identified by \code{SC3} as a heatmap}
\usage{
sc3_plot_de_genes.SCESet(object, k, p.val = 0.01, show_pdata = NULL)

\S4method{sc3_plot_de_genes}{SCESet}(object, k, p.val = 0.01,
  show_pdata = NULL)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}

\item{p.val}{significance threshold used for the DE genes}

\item{show_pdata}{a vector of colnames of the pData(object) table. Default is NULL.
If not NULL will add pData annotations to the columns of the output matrix}
}
\description{
\code{SC3} plots gene expression profiles of the 50 genes with the lowest p-values.
}

