% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_expression.SCESet}
\alias{sc3_plot_expression}
\alias{sc3_plot_expression,SCESet-method}
\alias{sc3_plot_expression.SCESet}
\title{Plot expression matrix used for SC3 clustering as a heatmap}
\usage{
sc3_plot_expression.SCESet(object, k, show_pdata = NULL)

\S4method{sc3_plot_expression}{SCESet}(object, k, show_pdata = NULL)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}

\item{show_pdata}{a vector of colnames of the pData(object) table. Default is NULL.
If not NULL will add pData annotations to the columns of the output matrix}
}
\description{
The expression panel represents the original input expression matrix 
(cells in columns and genes in rows) after cell and gene filters. 
Genes are clustered by kmeans with k = 100 (dendrogram on the left) and 
the heatmap represents the expression levels of the gene cluster centers 
after log2-scaling.
}

