% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_markers.SCESet}
\alias{sc3_plot_markers}
\alias{sc3_plot_markers,SCESet-method}
\alias{sc3_plot_markers.SCESet}
\title{Plot expression of marker genes of the clusters identified by SC3 as a heatmap}
\usage{
sc3_plot_markers.SCESet(object, k, auroc = 0.85, p.val = 0.01)

\S4method{sc3_plot_markers}{SCESet}(object, k, auroc = 0.85, p.val = 0.01)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}

\item{auroc}{area under the ROC curve}

\item{p.val}{significance threshold used for the DE genes}
}
\description{
To find marker genes, for each gene a binary classifier is constructed 
based on the mean cluster expression values. The classifier prediction 
is then calculated using the gene expression ranks. The area under the 
receiver operating characteristic (ROC) curve is used to quantify the accuracy 
of the prediction. A p-value is assigned to each gene by using the Wilcoxon 
signed rank test. By default the genes with the area under the ROC curve (AUROC) > 0.85 
and with the p-value < 0.01 are selected and the top 10 marker 
genes of each cluster are visualized in this heatmap.
}

