% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_silhouette.SCESet}
\alias{sc3_plot_silhouette}
\alias{sc3_plot_silhouette,SCESet-method}
\alias{sc3_plot_silhouette.SCESet}
\title{Plot silhouette indexes of the cells}
\usage{
sc3_plot_silhouette.SCESet(object, k)

\S4method{sc3_plot_silhouette}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}
}
\description{
A silhouette is a quantitative measure of the diagonality of the consensus 
matrix. An average silhouette width (shown at the bottom left of the silhouette 
plot) varies from 0 to 1, where 1 represents a perfectly block-diagonal 
consensus matrix and 0 represents a situation where there is no 
block-diagonal structure. The best clustering is achieved when the average 
silhouette width is close to 1.
}

