% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMethods.R
\docType{methods}
\name{sc3_plot_tsne.SCESet}
\alias{sc3_plot_tsne}
\alias{sc3_plot_tsne,SCESet-method}
\alias{sc3_plot_tsne.SCESet}
\title{Plot tSNE map of the cells and highlight SC3 clusters with colors}
\usage{
sc3_plot_tsne.SCESet(object, k,
  perplexity = floor(ncol(object@sc3$processed_dataset)/5), seed = 1234567)

\S4method{sc3_plot_tsne}{SCESet}(object, k,
  perplexity = floor(ncol(object@sc3$processed_dataset)/5), seed = 1234567)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{number of clusters}

\item{perplexity}{perplexity parameter used in \code{\link[Rtsne]{Rtsne}} for tSNE tranformation}

\item{seed}{random seed used for tSNE transformation}
}
\description{
\href{https://lvdmaaten.github.io/tsne/}{tSNE} (t-Distributed Stochastic 
Neighbor Embedding) method is used to map high-dimensional data to a 2D 
space while preserving local distances between cells. tSNE has become a 
very popular visualisation tool. SC3 imports the Rtsne function from the
\href{https://cran.r-project.org/web/packages/Rtsne/index.html}{Rtsne package} 
to perform the tSNE analysis. The colors on the plot correspond to the clusters 
identified by SC3. One of the most sensitive parameters in tSNE analysis is the
so-called perplexity. SC3 defines the default perplexity as N/5, where N is 
the number of cells.
}

