% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_run_svm.SCESet}
\alias{sc3_run_svm}
\alias{sc3_run_svm,SCESet-method}
\alias{sc3_run_svm.SCESet}
\title{Run SVM on training cells}
\usage{
sc3_run_svm.SCESet(object, k)

\S4method{sc3_run_svm}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{the number of clusters k for which the SVM should be run}
}
\value{
an object of 'SCESet' class
}
\description{
This function performs training of the SVM classifier on the training cells,
which indeces are  
contained in the svm_train_inds item of the object@sc3 slot. Then it 
predicts the labels of the remaining cells using the SVM classifier. Finally it
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item svm_result - contains labels of the cells predicted by the SVM ordered
  as the cells in the original dataset.
}
}

