% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_set_ks.SCESet}
\alias{sc3_set_ks}
\alias{sc3_set_ks,SCESet-method}
\alias{sc3_set_ks.SCESet}
\title{Sets a range of the number of clusters k used for SC3 clustering.}
\usage{
sc3_set_ks.SCESet(object, ks = NULL)

\S4method{sc3_set_ks}{SCESet}(object, ks = NULL)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{ks}{a continuous range of integers - the number of clusters k used for SC3 clustering.
Can also be a single integer.}
}
\description{
This function creates and populates the following items of the object@sc3 slot:
\itemize{
  \item ks - contains a range of the number of clusters k to be used by SC3
}
}

