% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_summarise_results.SCESet}
\alias{sc3_summarise_results}
\alias{sc3_summarise_results,SCESet-method}
\alias{sc3_summarise_results.SCESet}
\title{Summarise SC3 results}
\usage{
sc3_summarise_results.SCESet(object, k)

\S4method{sc3_summarise_results}{SCESet}(object, k)
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{k}{the number of clusters k for which the results should be summarised}
}
\value{
an object of 'SCESet' class
}
\description{
This function summarised all SC3 results into a single list and populates 
it to the following item of the object@sc3 slot:
\itemize{
  \item results - contains all SC3 results
}
}

