% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3min_export_results_xls}
\alias{sc3min_export_results_xls}
\alias{sc3min_export_results_xls.SingleCellExperiment}
\alias{sc3min_export_results_xls,SingleCellExperiment-method}
\title{Write \code{SC3min} results to Excel file}
\usage{
sc3min_export_results_xls.SingleCellExperiment(object, filename)

\S4method{sc3min_export_results_xls}{SingleCellExperiment}(object,
  filename = "sc3min_results.xls")
}
\arguments{
\item{object}{an object of \code{SingleCellExperiment} class}

\item{filename}{name of the excel file, to which the results will be written}
}
\description{
This function writes all \code{SC3min} results to an excel file.
}
