% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{splits}
\alias{splits}
\title{Split integral string into integral vector}
\usage{
splits(myString)
}
\arguments{
\item{myString}{An integral string.}
}
\description{
Similar to "strsplit" in R but only for integral strings and vectors.
}
\details{
To remove a note from check change this function using an answer here:
http://stackoverflow.com/questions/5607589/right-way-to-split-an-stdstring-into-a-vectorstring
}

