% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{support_vector_machines}
\alias{support_vector_machines}
\title{Support vector machines (SVM)}
\usage{
support_vector_machines(teach, study, kern)
}
\arguments{
\item{teach}{Expression matrix of teacher cells}

\item{study}{Expression matrix of study cells}

\item{kern}{Kernel to be used with SVM}
}
\value{
Classification of study cells
}
\description{
Train an SVM classifier on teacher cells and then
classify study cells usin the classifier
}

