% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{transformation}
\alias{transformation}
\title{Dimensionality reduction of a distance matrix}
\usage{
transformation(dists, method)
}
\arguments{
\item{dists}{A distance matrix}

\item{method}{Dimensionality reduction method: "pca", "spectral",
"spectral_reg" or "mds"}
}
\value{
A transformed distance matrix
}
\description{
Transform a distance matrix to a new basis
}

