% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-examples.R
\docType{data}
\name{treutlein}
\alias{treutlein}
\title{Single cell RNA-Seq data extracted from a publication by Treutlein et al.}
\format{An object of class \code{matrix} with 23271 rows and 80 columns.}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52583}

Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 80 cells and 5 clusters in this dataset.
}
\usage{
treutlein
}
\value{
blah blah
}
\description{
Single cell RNA-Seq data extracted from a publication by Treutlein et al.
}
\keyword{datasets}

