% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate a distance matrix}
\usage{
calculate_distance(data, method)
}
\arguments{
\item{data}{expression matrix}

\item{method}{one of the distance metrics: 'spearman', 'pearson', 'euclidean'}
}
\value{
distance matrix
}
\description{
Distance between the cells, i.e. columns, in the input expression matrix are
calculated using the Euclidean, Pearson and Spearman metrics to construct
distance matrices.
}
