% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{cell_filter}
\alias{cell_filter}
\title{Cell filter}
\usage{
cell_filter(data, cell.filter.genes)
}
\arguments{
\item{data}{expression matrix}

\item{cell.filter.genes}{minimum number of genes that must be expressed in
each cell, default is 2,000.}
}
\value{
Filtered expression matrix
}
\description{
The cell filter should be used if the quality of data is low, i.e. if one
suspects that some of the cells may be technical outliers with poor coverage.
The cell filter removes cells containing fewer than cell.filter.genes.
}

