% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consmx}
\alias{consmx}
\title{Consensus binary matrix computation}
\usage{
consmx(myString, res, length)
}
\arguments{
\item{myString}{Integral string corresponding to cluster labels.}

\item{res}{A square zero matrix, nrows = length. Will become
binary matrix.}

\item{length}{Integer, number of cells.}
}
\description{
Computes binary matrix given cluster labels
}

