% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{consmx}
\alias{consmx}
\title{Consensus matrix computation}
\usage{
consmx(dat)
}
\arguments{
\item{dat}{a matrix containing clustering solutions in columns}
}
\description{
Computes consensus matrix given cluster labels
}

