% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estkTW.R
\name{estkTW}
\alias{estkTW}
\title{Estimate the optimal k for k-means clustering}
\usage{
estkTW(dataset)
}
\arguments{
\item{dataset}{processed input expression matrix.}
}
\description{
The procedure filters and log-transforms the expression matrix before 
finding the eigenvalues of the sample covariance matrix. It will then return 
the number of significant eigenvalues according to the Tracy-Widom test.
}

