% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{gene_filter}
\alias{gene_filter}
\title{Gene filter}
\usage{
gene_filter(data, fraction = 0, reads.rare = 0, reads.ubiq = 0)
}
\arguments{
\item{data}{expression matrix}

\item{fraction}{fraction of cells (1 - X/100), default is 0.06.}

\item{reads.rare}{expression value threshold for genes that are expressed in
less than fraction*N cells (rare genes)}

\item{reads.ubiq}{expression value threshold for genes that are expressed in
more than (1-fraction)*N cells (ubiquitous genes)}
}
\value{
filtered expression matrix some genes were removed.
}
\description{
The gene filter removes genes that are either expressed or absent
(expression value is less than 2) in at least X % of cells.
The motivation for the gene filter is that ubiquitous and rare genes most
often are not informative for the clustering.
}

