% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{gene_filter}
\alias{gene_filter}
\title{Gene filter}
\usage{
gene_filter(data, fraction = 0.06, reads.rare = 2, reads.ubiq = 0)
}
\arguments{
\item{data}{input expression matrix}

\item{fraction}{fraction of cells (X/100).}

\item{reads.rare}{expression value threshold for genes that are expressed in
less than fraction*N cells (rare genes)}

\item{reads.ubiq}{expression value threshold for genes that are expressed in
more than (1-fraction)*N cells (ubiquitous genes)}
}
\value{
filtered expression matrix some genes were removed.
}
\description{
The gene filter removes genes/transcripts that are either expressed 
(expression value is more than 2) in less than X% of cells 
(rare genes/transcripts) or expressed (expression value is more than 0) 
in at least (100-X)% of cells (ubiquitous genes/transcripts).
}

