% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{get_data}
\alias{get_data}
\title{Import expression matrix}
\usage{
get_data(name)
}
\arguments{
\item{name}{name of an R object or a text file}
}
\value{
expression matrix
}
\description{
Import an input expression matrix.
}

