% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{get_de_genes}
\alias{get_de_genes}
\title{Find differentially expressed genes}
\usage{
get_de_genes(dataset, labels)
}
\arguments{
\item{dataset}{expression matrix}

\item{labels}{cell labels corresponding to the columns of the expression matrix}
}
\value{
a numeric vector containing the differentially expressed genes and 
correspoding p-values
}
\description{
If the cell labels are available this functions allows a user to calculate
differentially expressed genes manually.
}
\examples{
d <- get_de_genes(treutlein, colnames(treutlein))
head(d)

}

