% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Find marker genes}
\usage{
get_marker_genes(dataset, labels)
}
\arguments{
\item{dataset}{expression matrix}

\item{labels}{cell labels corresponding to the columns of the expression matrix.
Labels must be integers or character integers, e.g. 1, 2, 3 or '1', '2', '3' ect.}
}
\value{
data.frame containing the marker genes
}
\description{
If the cell labels are available this functions allows a user to calculate
marker genes manually.
}
\examples{
d <- get_marker_genes(treutlein, colnames(treutlein))
head(d)

}

