% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFunctions.R
\name{get_marker_genes}
\alias{get_marker_genes}
\title{Calculate marker genes}
\usage{
get_marker_genes(dataset, labels)
}
\arguments{
\item{dataset}{expression matrix}

\item{labels}{cell labels corresponding clusters}
}
\value{
data.frame containing the marker genes, corresponding cluster indexes
and adjusted \code{p-value}s
}
\description{
Find marker genes in the dataset. The \code{\link{get_auroc}} is used to calculate
marker values for each gene.
}
\examples{
d <- get_marker_genes(treutlein[1:10,], colnames(treutlein))
d

}
