% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_processed_dataset}
\alias{get_processed_dataset}
\title{Get processed dataset used by SC3 from the default scater slots}
\usage{
get_processed_dataset(object, exprs_values = "counts")
}
\arguments{
\item{object}{an object of 'SCESet' class}

\item{exprs_values}{character string 
indicating which values should be used
as the expression values for SC3 clustering. Valid arguments are \code{'tpm'}
(default; transcripts per million), \code{'norm_tpm'} (normalised TPM
values), \code{'fpkm'} (FPKM values), \code{'norm_fpkm'} (normalised FPKM
values), \code{'counts'} (counts for each feature), \code{'norm_counts'},
\code{'cpm'} (counts-per-million), \code{'norm_cpm'} (normalised
counts-per-million), \code{'exprs'} (whatever is in the \code{'exprs'} slot
of the \code{SCESet} object; default), \code{'norm_exprs'} (normalised
expression values) or \code{'stand_exprs'} (standardised expression values)
or any other named element of the \code{assayData} slot of the \code{SCESet}
object that can be accessed with the \code{get_exprs} function.}
}
\description{
Takes data from the 'exprs_values' slot, applies gene filter and log
transformation.
}

