% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{get_processed_dataset}
\alias{get_processed_dataset}
\title{Get processed dataset used by SC3 from the default scater slots}
\usage{
get_processed_dataset(object, exprs_values = "counts")
}
\arguments{
\item{object}{}

\item{exprs_values}{}
}
\description{
Takes data from the 'exprs_values' slot, applies gene filter and log
transformation.
}

