% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{norm_laplacian}
\alias{norm_laplacian}
\title{Graph Laplacian calculation}
\usage{
norm_laplacian(A)
}
\arguments{
\item{A}{symmetric matrix}
}
\description{
Calculate graph Laplacian of a symmetrix matrix
}
