% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{norm_laplacian}
\alias{norm_laplacian}
\title{Graph Laplacian calculation}
\usage{
norm_laplacian(x, tau)
}
\arguments{
\item{x}{adjacency/distance matrix}

\item{tau}{regularization term}
}
\value{
graph Laplacian of the adjacency/distance matrix
}
\description{
Calculate graph Laplacian of a distance matrix
}

