% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3.R
\name{sc3}
\alias{sc3}
\title{SC3 main function}
\usage{
sc3(dataset, ks, gene.filter = TRUE, gene.filter.fraction = 0.06,
  gene.reads.rare = 2, gene.reads.ubiq = 0, log.scale = TRUE,
  d.region.min = 0.04, d.region.max = 0.07, k.means.iter.max = 1e+09,
  k.means.nstart = 1000, show.original.labels = FALSE,
  svm.num.cells = NULL, svm.train.inds = NULL, n.cores = NULL, seed = 1)
}
\arguments{
\item{dataset}{input expression matrix.}

\item{ks}{a range of the number of clusters that needs to be tested.
k.min is the minimum number of clusters (default is 3). k.max is the maximum
number of clusters (default is 7).}

\item{gene.filter}{defines whether to perform gene filtering or not. Boolean,
default is TRUE.}

\item{gene.filter.fraction}{fraction of cells (1 - X/100), default is 0.06.
The gene filter removes genes that are either expressed or absent
(expression value is less than 2) in at least X % of cells.
The motivation for the gene filter is that ubiquitous and rare genes most
often are not informative for the clustering.}

\item{gene.reads.rare}{expression value threshold for genes that are expressed in
less than gene.filter.fraction*N cells (rare genes)}

\item{gene.reads.ubiq}{expression value threshold for genes that are expressed in
more than (1-fraction)*N cells (ubiquitous genes)}

\item{log.scale}{defines whether to perform log2 scaling or not. Boolean,
default is TRUE.}

\item{d.region.min}{the lower boundary of the optimum region of d, 
default is 0.04.}

\item{d.region.max}{the upper boundary of the optimum region of d, 
default is 0.07.}

\item{k.means.iter.max}{iter.max parameter used by kmeans() function. Default is 1e+09.}

\item{k.means.nstart}{nstart parameter used by kmeans() function. Default is 1000.}

\item{show.original.labels}{if cell labels in the dataset are not unique,
but represent clusters expected from the experiment, they can be visualised
by setting this parameter to TRUE. The default is FALSE.}

\item{svm.num.cells}{number of training cells to be used for SVM prediction. 
The default is NULL. If the svm parameter is TRUE and svn.num.cells is not provided,
then the svm.train.inds parameter is checked.}

\item{svm.train.inds}{a numeric vector defining indeces of training cells that should be used for SVM training. 
The default is NULL. If the svm parameter is TRUE and svn.num.cells and svm.train.inds are not provided,
then the defaults of SC3 will be used: if number of cells is more than 5000,
then svn.num.cells = 1000, otherwise svn.num.cells = 20 percent of the total number of cells}

\item{n.cores}{defines the number
of cores to be used on the user's machine. Default is NA.}

\item{seed}{sets seed for the random number generator, default is 1.
Can be used to check the stability of clustering results: if the results are 
the same after changing the seed several time, then the clustering solution 
is stable.}
}
\value{
Opens a browser window with an interactive shine app and visualize
all precomputed clusterings.
}
\description{
Run SC3 clustering pipeline and starts an interactive session in a web browser.
}
\examples{
sc3(treutlein, 3:7, n.cores = 2)

}

