% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3.R
\name{sc3}
\alias{sc3}
\title{SC3 main function}
\usage{
sc3(filename, ks = 3:7, cell.filter = FALSE, cell.filter.genes = 2000,
  gene.filter.fraction = 0.06, d.region.min = 0.04, d.region.max = 0.07,
  chisq.quantile = 0.9999, auroc.threshold = 0.85, interactivity = TRUE,
  show.original.labels = FALSE, svm.num.cells = 1000,
  use.max.cores = TRUE, n.cores = 1, seed = 1)
}
\arguments{
\item{filename}{either an R matrix / data.frame / data.table object OR a
path to your input file containing an input expression matrix.}

\item{ks}{a range of the number of clusters that needs to be tested.
k.min is the minimum number of clusters (default is 3). k.max is the maximum
number of clusters (default is 7).}

\item{cell.filter}{defines whether to filter cells that express less than
cell.filter.genes genes (lowly expressed cells). By default it is FALSE.
The cell filter should be used if the quality of data is low, i.e. if one
suspects that some of the cells may be technical outliers with poor coverage.
Filtering of lowly expressed cells usually improves clustering.}

\item{cell.filter.genes}{if cell.filter is used then this parameter defines
the minimum number of genes that have to be expressed in each cell
(expression value > 1e-2). If there are fewer, the cell will be
removed from the analysis. The default is 2000.}

\item{gene.filter.fraction}{fraction of cells (1 - X/100), default is 0.06.
The gene filter removes genes that are either expressed or absent
(expression value is less than 2) in at least X % of cells.
The motivation for the gene filter is that ubiquitous and rare genes most
often are not informative for the clustering.}

\item{d.region.min}{the lower boundary of the optimum region of d, 
default is 0.04.}

\item{d.region.max}{the upper boundary of the optimum region of d, 
default is 0.07.}

\item{chisq.quantile}{a threshold used for cell outliers detection, 
default is 0.9999.}

\item{auroc.threshold}{a threshold used for marker genes detection,
default is 0.85}

\item{interactivity}{defines whether a browser interactive window should be
open after all computation is done. By default it is TRUE. This option can
be used to separate clustering calculations from visualisation,
e.g. long and time-consuming clustering of really big datasets can be run
on a farm cluster and visualisations can be done using a personal
laptop afterwards. If interactivity is FALSE then all clustering results
will be saved as "sc3.interactive.arg" list. To run interactive visulisation with
the precomputed clustering results please use
sc3_interactive(sc3.interactive.arg).}

\item{show.original.labels}{if cell labels in the dataset are not unique,
but represent clusters expected from the experiment, they can be visualised
by setting this parameter to TRUE. The default is FALSE.}

\item{svm.num.cells}{if number of cells in the dataset is greater than this
parameter, then an SVM prediction will be used. The default is 1000.}

\item{use.max.cores}{defines whether to use maximum available cores on the
user's machine. Logical, default is TRUE.}

\item{n.cores}{if use.max.cores is FALSE, this parameter defines the number
of cores to be used on the user's machine. Default is 1.}

\item{seed}{sets seed for the random number generator, default is 1.
Can be used to check the stability of clustering results: if the results are 
the same after changing the seed several time, then the clustering solution 
is stable.}
}
\value{
Opens a browser window with an interactive shine app and visualize
all precomputed clusterings.
}
\description{
Run SC3 clustering pipeline and starts an interactive session in a web browser.
}
\examples{
sc3(treutlein, 3:7, interactivity = FALSE, use.max.cores = FALSE)

}

