% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_calc_dists}
\alias{sc3_calc_dists}
\alias{sc3_calc_dists.SCESet}
\alias{sc3_calc_dists,}
\alias{sc3_calc_dists,SCESet-method}
\alias{sc3_calc_dists,SCESet-method}
\alias{sc3_calc_dists}
\title{Calculate distances between the cells.}
\usage{
sc3_calc_dists.SCESet(object)

\S4method{sc3_calc_dists}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function calculates distances between the cells contained in 
the processed_dataset item of the \code{sc3} slot of the \code{SCESet} object. It then
creates and populates the following items of the \code{sc3} slot:
\itemize{
  \item \code{distances} - contains a list of distance matrices corresponding to
  Euclidean, Pearson and Spearman distances.
}
Please note that by default the "exprs" slot of the input \code{scater} object 
is used for the SC3 analysis. If the scater object has been created in a standard
way then the expression values in the "exprs" slot will be automatically log-transformed.
If you have overwritten the "exprs" slot manually, please make sure that the values
in the "exprs" slot are log-transformed before running the SC3 analysis. SC3
assumes that the data is log-transformed by default.
}
