% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sc3_interactive}
\alias{sc3_interactive}
\title{SC3 interactive function}
\usage{
sc3_interactive(input.param)
}
\arguments{
\item{input.param}{parameters precomputed by sc3() with interactivity = FALSE (sc3.interactive.arg).}
}
\value{
Opens a browser window with an interactive shiny app and visualize
all precomputed clusterings.
}
\description{
Runs interactive session of SC3 based on precomputed objects
}

