% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyMethods.R
\docType{methods}
\name{sc3_interactive.SCESet}
\alias{sc3_interactive}
\alias{sc3_interactive,SCESet-method}
\alias{sc3_interactive.SCESet}
\title{Open SC3 results in an interactive session in a web browser}
\usage{
sc3_interactive.SCESet(object)

\S4method{sc3_interactive}{SCESet}(object)
}
\arguments{
\item{object}{an object of "SCESet" class}
}
\value{
Opens a browser window with an interactive shiny app and visualize
all precomputed clusterings.
}
\description{
Runs interactive session of SC3 based on precomputed objects
}

