% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_kmeans}
\alias{sc3_kmeans}
\alias{sc3_kmeans,}
\alias{sc3_kmeans,SCESet-method}
\alias{sc3_kmeans.SCESet}
\title{\code{kmeans} clustering of cells.}
\usage{
sc3_kmeans.SCESet(object)

\S4method{sc3_kmeans}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function performs \code{\link[stats]{kmeans}} clustering of the matrices 
contained in the \code{transformations} item of the \code{sc3} slot of the \code{SCESet} object. It then
creates and populates the following items of the \code{sc3} slot:
\itemize{
  \item \code{kmeans} - contains a list of kmeans clusterings.
}
Additionally, it also removes the previously calculated \code{transformations} from
the \code{sc3} slot, as they are not needed for further analysis.
}
\details{
See \code{\link{sc3_prepare}} for the default clustering parameters.
}

