% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\docType{methods}
\name{sc3_kmeans.SCESet}
\alias{sc3_kmeans}
\alias{sc3_kmeans,SCESet-method}
\alias{sc3_kmeans.SCESet}
\title{kmeans clustering of the transformed distance matrices.}
\usage{
sc3_kmeans.SCESet(object)

\S4method{sc3_kmeans}{SCESet}(object)
}
\arguments{
\item{object}{an object of 'SCESet' class}
}
\value{
an object of 'SCESet' class
}
\description{
This function performs kmeans clustering of the transformed distance matrices 
contained in the transformations item of the object@sc3 slot. It then
creates and populates the following items of the object@sc3 slot:
\itemize{
  \item kmeans - contains a list of kmeans clusterings.
}
}
\details{
By default the nstart parameter passed to \link[stats]{kmeans} is defined
in \code{\link{sc3_prepare.SCESet}}, is set to 1000 and written to 
kmeans_nstart item of the object@sc3 slot.
If the number of cells in the dataset is more than 2000, this parameter is 
set to 50. A user can also manually define this parameter by changing the 
value of the kmeans_nstart item of the object@sc3 slot.
}

