% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc3-funcs.R
\name{support_vector_machines}
\alias{support_vector_machines}
\title{Run support vector machines (SVM) prediction}
\usage{
support_vector_machines(train, study, kern)
}
\arguments{
\item{train}{expression matrix with training cells}

\item{study}{expression matrix with study cells}

\item{kern}{kernel to be used with SVM}
}
\value{
classification of study cells
}
\description{
Train an SVM classifier on training cells and then
classify study cells using the classifier.
}

