% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{support_vector_machines}
\alias{support_vector_machines}
\title{Run support vector machines (\code{SVM}) prediction}
\usage{
support_vector_machines(train, study, kern)
}
\arguments{
\item{train}{training dataset with colnames, corresponding to training labels}

\item{study}{study dataset}

\item{kern}{kernel to be used with SVM}
}
\value{
classification of the study dataset
}
\description{
Train an \code{SVM} classifier on a training dataset (\code{train}) and then
classify a study dataset (\code{study}) using the classifier.
}

