% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreFunctions.R
\name{transformation}
\alias{transformation}
\title{Distance matrix transformation}
\usage{
transformation(dists, method, n_dim)
}
\arguments{
\item{dists}{distance matrix}

\item{method}{transformation method: either 'pca' or
'laplacian'}

\item{n_dim}{number, calculate only the first n_dim top dimensions}
}
\value{
transformed distance matrix
}
\description{
All distance matrices are transformed using either principal component 
analysis (PCA) or by calculating the 
eigenvectors of the graph Laplacian (Spectral). 
The columns of the resulting matrices are then sorted in 
descending order by their corresponding eigenvalues.
}
